requirements.txt
certifi>=2016.2.28
cloup>=0.14.1
humanize>=0.5.1
python-slugify>=1.2.1
tqdm>=4.1.0
urllib3>=1.19.1
